requirements.txt
datafusion>=42.0
streamlit>=1.35
pandas>=2.0
polars>=1.0
duckdb>=1.0
pyarrow>=15
azure-storage-blob>=12
azure-storage-file-share>=12
azure-identity>=1.15
python-dotenv>=1.0
plotly>=5.22
openpyxl>=3.1
